Modularize build.py
