+++ resources/validation.py
@@ -1,5 +1,6 @@
 import subprocess
-from resources import build, sys_patch_helpers
+from resources import sys_patch_helpers
+from resources.build import build
 from data import example_data, model_array, sys_patch_dict, os_data
 from pathlib import Path
